file hpbar, max_health can be any int not just 4
