+++ scripts/ball.py
@@ -6,12 +6,12 @@ from constant_values import FPS, LEFT, RIGHT, NONE
 
 class Ball(pygame.sprite.Sprite):
     DIAMETER = 20
-    DECELERATION = 0.995
+    DECELERATION = 1
 
     def __init__(self, x, y):
         pygame.sprite.Sprite.__init__(self)
 
-        ball_img = pygame.image.load(os.path.join('Assets', 'balls', 'basket-ball.png')).convert_alpha()
+        ball_img = pygame.image.load(os.path.join('scripts/Assets', 'balls', 'basket-ball.png')).convert_alpha()
         ball_img_scaled = pygame.transform.scale(ball_img, (self.DIAMETER, self.DIAMETER))
         self.image = ball_img_scaled
         self.rect = self.image.get_rect(center=(x, y))
@@ -89,7 +89,7 @@ class Ball(pygame.sprite.Sprite):
 class Cue(pygame.sprite.Sprite):
     def __init__(self, pos):
         pygame.sprite.Sprite.__init__(self)
-        self.cue_original_img = pygame.image.load(os.path.join('Assets', 'players', 'cue.png')).convert_alpha()
+        self.cue_original_img = pygame.image.load(os.path.join('scripts/Assets', 'players', 'cue.png')).convert_alpha()
         self.angle = 0
         self.image = pygame.transform.rotate(self.cue_original_img, self.angle)
         self.rect = self.image.get_rect(center=pos)
